Update README.md
The Online Bus Pass Application System is a modern, web-based platform designed to simplify the process of applying for and managing bus passes. Whether you're a commuter, a student, or a regular traveler, our system offers an efficient and user-friendly way to access, apply for, and renew bus passes, all from the comfort of your home.